-Adds SignUpScreenScaffold in order to have access to preview composable -Refactors SignUpScreen to use SignUpScreenScaffold and remove unused imports -Adds package for Screens and Signup in auth package for organization -Removes unused imports in MainActivity
